notebooks/advance-track/requirements.txt
numpy==1.26.4
pandas==2.2.3
matplotlib==3.7.0
seaborn==0.12.2
scikit-learn==1.6.1
tensorflow==2.15.0
torch==2.5.1
sympy==1.13.1
pykan==0.2.8
ipykernel
pyyaml
tqdm
pyarrow
ipywidgets